+++ Logic_ex.c
@@ -1,7 +1,7 @@
 /*!\file Logic_ex.c
 ** \author SMFSW
 ** \copyright MIT (c) 2017-2020, SMFSW
-** \brief Simple extension for logic variable handling
+** \brief Extensions for logic variable handling
 **/
 /****************************************************************/
 #include <string.h>